Move create_minipreview functions to icon.c
Just move the functions. This functions are related to the icon, so
they should be in this file.
